+++ lib/src/recovery.h
@@ -77,7 +77,7 @@ on_pkt_sent(struct q_conn * const c, struct w_iov * const v);
 
 extern void __attribute__((nonnull)) on_ack_rx_1(struct q_conn * const c,
                                                  const uint64_t ack,
-                                                 const uint16_t ack_delay);
+                                                 const uint64_t ack_delay);
 
 extern void __attribute__((nonnull)) on_ack_rx_2(struct q_conn * const c);
